# backend/docs/swagger.yaml
basePath: /api
definitions:
  dto.LoginRequest:
    properties:
      password:
        minLength: 6
        type: string
      username:
        maxLength: 50
        minLength: 3
        type: string
    required:
    - password
    - username
    type: object
  dto.LoginResponse:
    properties:
      token:
        type: string
      user:
        properties:
          id:
            type: integer
          role:
            type: string
          username:
            type: string
        type: object
    type: object
  dto.MeResponse:
    properties:
      created_at:
        type: string
      email:
        type: string
      full_name:
        type: string
      id:
        type: integer
      role:
        type: string
      username:
        type: string
    type: object
  dto.RegisterRequest:
    properties:
      email:
        type: string
      full_name:
        type: string
      password:
        minLength: 8
        type: string
      role:
        enum:
        - admin
        - editor
        type: string
      username:
        maxLength: 50
        minLength: 3
        type: string
    required:
    - email
    - password
    - role
    - username
    type: object
  middlewares.ErrorResponse:
    properties:
      error:
        type: string
    type: object
  project.CreateProjectRequest:
    properties:
      created_by:
        enum:
        - jeyshid
        - lodge
        type: string
      description:
        type: string
      display_order:
        type: integer
      employer:
        type: string
      ended_at:
        type: string
      ended_at_jalali:
        type: string
      full_name:
        maxLength: 255
        type: string
      is_visible:
        type: boolean
      slug:
        maxLength: 100
        type: string
      started_at:
        type: string
      started_at_jalali:
        type: string
    required:
    - created_by
    - full_name
    - slug
    type: object
  project.ImageResponse:
    properties:
      alt_text:
        type: string
      created_at:
        type: string
      id:
        type: integer
      image_url:
        type: string
      project_id:
        type: integer
      sort_order:
        type: integer
    type: object
  project.PaginatedProjectsResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/project.ProjectResponse'
        type: array
      page:
        type: integer
      per_page:
        type: integer
      total:
        type: integer
      total_pages:
        type: integer
    type: object
  project.ProjectResponse:
    properties:
      created_at:
        type: string
      created_by:
        type: string
      description:
        type: string
      display_order:
        type: integer
      employer:
        type: string
      ended_at:
        type: string
      ended_at_jalali:
        type: string
      full_name:
        type: string
      id:
        type: integer
      is_visible:
        type: boolean
      slug:
        type: string
      started_at:
        type: string
      started_at_jalali:
        type: string
      updated_at:
        type: string
    type: object
  project.UpdateProjectRequest:
    properties:
      description:
        type: string
      display_order:
        type: integer
      employer:
        type: string
      ended_at:
        type: string
      ended_at_jalali:
        type: string
      full_name:
        type: string
      is_visible:
        type: boolean
      started_at:
        type: string
      started_at_jalali:
        type: string
    type: object
  user.CreateUserRequest:
    properties:
      email:
        type: string
      full_name:
        type: string
      password:
        minLength: 8
        type: string
      role:
        enum:
        - admin
        - editor
        type: string
      username:
        maxLength: 50
        minLength: 3
        type: string
    required:
    - email
    - password
    - role
    - username
    type: object
  user.UpdateUserRequest:
    properties:
      email:
        type: string
      full_name:
        type: string
      password:
        minLength: 8
        type: string
      role:
        enum:
        - admin
        - editor
        type: string
    type: object
  user.UserResponse:
    properties:
      created_at:
        type: string
      email:
        type: string
      full_name:
        type: string
      id:
        type: integer
      role:
        type: string
      updated_at:
        type: string
      username:
        type: string
    type: object
host: localhost:8080
info:
  contact:
    email: jeyshid@example.com
    name: Jeyshid
  description: Personal portfolio & project management API
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: https://partojshid.ir/terms
  title: Partojshid API
  version: "1.0"
paths:
  /auth/login:
    post:
      consumes:
      - application/json
      description: Authenticate and receive JWT token
      parameters:
      - description: Login credentials
        in: body
        name: credentials
        required: true
        schema:
          $ref: '#/definitions/dto.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.LoginResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/middlewares.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/middlewares.ErrorResponse'
      summary: Login user
      tags:
      - Auth
  /auth/me:
    get:
      description: Returns authenticated user's profile
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.MeResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/middlewares.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get current user
      tags:
      - Auth
  /auth/register:
    post:
      consumes:
      - application/json
      parameters:
      - description: User registration data
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/dto.RegisterRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/user.UserResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/middlewares.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/middlewares.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/middlewares.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Register new user (admin only)
      tags:
      - Auth
  /projects:
    get:
      consumes:
      - application/json
      description: Get paginated list of projects. Public sees only visible ones.
        Use ?editor=1 with admin/editor token to see hidden projects.
      parameters:
      - default: 1
        description: Page number
        in: query
        minimum: 1
        name: page
        type: integer
      - default: 10
        description: Items per page
        in: query
        maximum: 100
        minimum: 1
        name: per_page
        type: integer
      - description: Show hidden projects (admin/editor only)
        in: query
        name: editor
        type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/project.PaginatedProjectsResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/middlewares.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/middlewares.ErrorResponse'
      summary: List projects
      tags:
      - Projects
    post:
      consumes:
      - application/json
      description: Only editors and admins can create projects
      parameters:
      - description: Project data
        in: body
        name: project
        required: true
        schema:
          $ref: '#/definitions/project.CreateProjectRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/project.ProjectResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/middlewares.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/middlewares.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/middlewares.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Create a new project
      tags:
      - Projects (Admin)
  /projects/{id}:
    delete:
      description: Only editors and admins can delete
      parameters:
      - description: Project ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "204":
          description: No Content
          schema:
            type: string
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/middlewares.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/middlewares.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/middlewares.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Delete a project
      tags:
      - Projects (Admin)
    get:
      consumes:
      - application/json
      description: Retrieve a single project by ID. Hidden projects return 404 for
        public users.
      parameters:
      - description: Project ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/project.ProjectResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/middlewares.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/middlewares.ErrorResponse'
      summary: Get a project
      tags:
      - Projects
    put:
      consumes:
      - application/json
      description: Only editors and admins can update
      parameters:
      - description: Project ID
        in: path
        name: id
        required: true
        type: integer
      - description: Updated fields
        in: body
        name: project
        required: true
        schema:
          $ref: '#/definitions/project.UpdateProjectRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/project.ProjectResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/middlewares.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/middlewares.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/middlewares.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update a project
      tags:
      - Projects (Admin)
  /projects/{id}/images:
    get:
      description: Get all images for a project
      parameters:
      - description: Project ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/project.ImageResponse'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/middlewares.ErrorResponse'
      summary: List project images
      tags:
      - Projects
    post:
      consumes:
      - multipart/form-data
      description: Upload a single image (max 5MB, jpg/png/webp)
      parameters:
      - description: Project ID
        in: path
        name: id
        required: true
        type: integer
      - description: Image file
        in: formData
        name: image
        required: true
        type: file
      - description: Alt text (max 255 chars)
        in: formData
        name: alt_text
        type: string
      - description: Sort order
        in: formData
        minimum: 0
        name: sort_order
        type: integer
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/project.ImageResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/middlewares.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/middlewares.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/middlewares.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Upload image to project
      tags:
      - Projects (Admin)
  /projects/{id}/images/{image_id}:
    delete:
      description: Permanently delete an image (admin/editor only)
      parameters:
      - description: Project ID
        in: path
        name: id
        required: true
        type: integer
      - description: Image ID
        in: path
        name: image_id
        required: true
        type: integer
      responses:
        "204":
          description: No Content
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/middlewares.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/middlewares.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Delete project image
      tags:
      - Projects (Admin)
  /users:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/user.UserResponse'
            type: array
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/middlewares.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/middlewares.ErrorResponse'
      security:
      - BearerAuth: []
      summary: List all users (admin only)
      tags:
      - Users (Admin)
    post:
      consumes:
      - application/json
      parameters:
      - description: User data
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/user.CreateUserRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/user.UserResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/middlewares.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/middlewares.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/middlewares.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Create new user (admin only)
      tags:
      - Users (Admin)
  /users/{id}:
    delete:
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "204":
          description: No Content
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/middlewares.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/middlewares.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Delete user (admin only)
      tags:
      - Users (Admin)
    get:
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/user.UserResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/middlewares.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get user by ID (admin only)
      tags:
      - Users (Admin)
    put:
      consumes:
      - application/json
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      - description: Updated fields
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/user.UpdateUserRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/user.UserResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/middlewares.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/middlewares.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/middlewares.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update user (admin only)
      tags:
      - Users (Admin)
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"